daily scrum Larisa
We should communicate on the details of how to work in the following days to avoid the missunderstandings. I will work now on a slight change of design for the Reports part. The pages See budget, See expenses and See income will be incorporated in reports as well as the add income and expenses will be accessible form reports probably on right click.